Refactor to use useDeleteTask
