+++ frontend/src/components/project/taskDetails/mobile/TaskCreateMobile.jsx
@@ -4,7 +4,7 @@ import { useLocation, useOutletContext } from "react-router-dom"
 import TaskCommonDetailMobile from "@components/project/taskDetails/mobile/TaskCommonDetailMobile"
 
 const TaskCreateMobile = ({ closeCreate }) => {
-    const [projectID, projectType, color] = useOutletContext()
+    const [_, projectType, color] = useOutletContext()
     const { state } = useLocation()
 
     const [newTask, setNewTask] = useState({
@@ -28,7 +28,6 @@ const TaskCreateMobile = ({ closeCreate }) => {
         <TaskCommonDetailMobile
             newTask={newTask}
             setNewTask={setNewTask}
-            projectID={projectID}
             projectType={projectType}
             color={color}
             onClose={closeCreate}
